fix: loading list componentify
